requirements.txt
# libs do projeto
#rode o comando: pip install -r requirements.txt
customtkinter
Pillow
Jinja2
weasyprint
pymupdf
matplotlib
